Lexer

Description

Documentation

between :: Char -> Char -> Parser a -> Parser a

Runs a parser between two delimiters.

brackets :: Parser a -> Parser a

Runs a parser between two brackets { }.

char :: Parser Char

Parses one character, returning the character and the rest of the unparsed string.

digit :: Parser Char

Parses a digit.

digits :: Parser String

Parses multiple digits, returning them as a string.

end :: Parser ()

Parses the end of an instruction, denoted by a semicolon. Discards result.

ident :: String -> Parser ()

Parses an identifier. Discards the result.

keyword :: String -> Parser ()

Parses an identifier followed by a semicolon. Discards the result.

letter :: Parser Char

Parses a letter, returning the letter.

letters :: Parser String

Parses multiple letters, returning them as a string.

lower :: Parser Char

Parses a lowercase letter.

optional :: Parser String -> Parser String

Parses a string if it is found, returning the empty string otherwise.

parens :: Parser a -> Parser a

Runs a parser between parentheses ( ).

semicolon :: Parser ()

Parses a semicolon. Discards the result.

skipUntil :: Parser a -> Parser String

Parses characters as long as a given condition holds true.

spot :: (Char -> Bool) -> Parser Char

Parses a character satisfying a given predicate.

string :: String -> Parser String

Parses a string.

token :: Char -> Parser ()

Parses a given character. Discards the result.

upper :: Parser Char

Parses an uppercase letter.

preprocess :: String -> String

Preprocesses a file for parsing, removing all whitespace.

bool :: Parser Exp

Tokenizes a boolean expression.

binaryExpr :: Parser Exp

Tokenizes a binary expression.

color :: Parser Color

Tokenizes a color.

constant :: Parser Exp

Tokenizes a constantvariablerobot expression.

expr :: Parser Exp

Tokenizes any expression.

number :: Parser Double

Tokenizes a number expression.

robotDirection :: Parser Direction

Tokenizes an MBot direction.

robotLed :: Parser Led

Tokenizes an MBot LED identifier.

unaryExpr :: Parser Exp

Tokenizes a unary expression.