MBot

Description

Programming the mBot

With this library it is possible to control the mBot robot from within Haskell over 2.4ghz wireless. The mBot itself needs to contain the standard firmware otherwise the library will not behave as expected. There is support for steering the motors and leds and for reading the linesensor and the ultrasonic sensor. An small example program is shown below, for more information about the individual functions take a look at the api documentation below.

import MBot

main =  do
  putStrLn "My first mBot program in Haskell !"
  -- Open the connection with the mMbot
  d <- openMBot
  putStrLn "Opened a connection with the mBot"
  -- Turn on led 1 of the mBot and set the RGB value to (0,100,0)
  sendCommand d $ setRGB 1 0 100 0
  putStrLn "Look at all the pretty colors !"
  -- Turn on led 2 of the mBot and set the RGB value to (100,0,0)
  sendCommand d $ setRGB 2 100 0 0
  -- close the connection with the mBot
  closeMBot d

Documentation

openMBot

Arguments

:: IO Device
gives back the connection with the mBot

Opens a connection with the mBot

closeMBot :: Device -> IO ()

Close the connection with the mBot

sendCommand

Arguments

:: Device
An open Device connection

-> Command
The command to send

-> IO ()
There is no return value

Sends a mBot command over the HID device

readUltraSonic :: Device -> IO Float

Read out the status of the ultrasonic line follower

readLineFollower :: Device -> IO Line

Read out the status of line follower sensor

goAhead :: Device -> IO ()

Start both motors so that the robot moves fowards

goBackwards :: Device -> IO ()

Start both motors so that the robot moves backwards

goLeft :: Device -> IO ()

Start the motors let the mBot turn left

goRight :: Device -> IO ()

Start the motors so that the robots turns right

stop :: Device -> IO ()

Stop both motors

setRGB :: Int -> Int -> Int -> Int -> Command

Create an mBot command to turn on the led on a particular rgb value

setMotor :: Int -> Int -> Int -> Command

data Line

The line sensor consists of two sensors which are able to detect either a black or a white surface. Therefore there are four different states to represent the state of the line sensor

Constructors

LEFTB
Left sensor reads black right sensor reads white

RIGHTB
Right sensor reads black left sensor reads white

BOTHB
Both the left and right sensor observe a black surface

BOTHW
Both the left and right sensor observe a white surface

data Command

Type of mBot commands the constructor is not exported.