MBotPlus

Description

Documentation

data Direction

Driving directions of the motors.

Constructors

DirForward
Drive forward

DirLeft
Turn left

DirRight
Turn right

Brake

DirBackward
Drive backwards

DirBackwardLeft
Drive backwards and turn left

DirBackwardRight
Drive backwards and turn right

data Led

Led identifiers.

Constructors

LeftLed
The left led

RightLed
The right led

data Motor

Motor identifiers.

Constructors

LeftMotor
The left motor

RightMotor
The right motor

backwardsLeft :: Device -> IO ()

Motor command: drive backwards and turn left.

backwardsRight :: Device -> IO ()

Motor command: drive backwards and turn right.

close :: Device -> IO ()

Close the connection to the MBot.

connect :: IO Device

Open a connection to the MBot.

command :: Device -> Command -> IO ()

Sends a comment to the MBot.

ledId :: Led -> Int

Retrieves a LED int by its identifier.

led :: Device -> Led -> Double -> Double -> Double -> IO ()

Sets the color on a given LED.

lineDouble :: Line -> Double

Converts the linesensor value to a Double. BOTHW = 0, LEFTB = 1, RIGHTB = 2, BOTHB = 3

lineSensor :: Device -> IO Double

Gets the value of the linesensor as a Double.

motorDirection :: Direction -> Device -> IO ()

Gets the appropriate motor instructions for a given Direction.

motorId :: Motor -> Int

Gets the motor int by its identifier.

ultrason :: Device -> IO Double

Gets the value of the ultrasonic sensor as a Double.