Utils

Description

Documentation

boolDouble :: Bool -> Double

Converts a Boolean to a Double value.

doubleBool :: Double -> Bool

Converts a Double to a Boolean value. Every number is converted to True except for zero.

doubleInt :: Double -> Int

Converts a Double to an Integer.

floatDouble :: Float -> Double

Converts a floating point number to a Double.

index :: Eq a => [a] -> a -> Int

Retrieves the index of an element in a list, erroring if not found.

intDouble :: Int -> Double

Converts an Integer to a Double.

mapLookup :: Ord a => [(a, b)] -> a -> Maybe b

Retrieves the value corresponding to a key in a list of tuples.